implement download progress bar, refactor using service
