+++ UI/Contracts/IGitHubService.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Threading.Tasks;
 using MuseDashModToolsUI.Models;
 
@@ -8,6 +9,6 @@ public interface IGitHubService
 {
     Task<List<Mod>> GetModsAsync();
     Task DownloadModAsync(string link, string path);
-    Task DownloadMelonLoader(string path, double downloadProgress, bool finished);
+    Task DownloadMelonLoader(string path, IProgress<int> downloadProgress);
     void CheckUpdates();
 }
