+++ UI/Contracts/ISettingService.cs
@@ -1,0 +1,11 @@
+﻿using System.Threading.Tasks;
+using MuseDashModToolsUI.Models;
+
+namespace MuseDashModToolsUI.Contracts;
+
+public interface ISettingService
+{
+    public Setting Settings { get; set; }
+    Task InitializeSettings();
+    Task OnChoosePath();
+}
