+++ UI/Models/Setting.cs
@@ -1,10 +1,12 @@
-﻿using MuseDashModToolsUI.Contracts;
+﻿using System.IO;
+using System.Text.Json.Serialization;
 
 namespace MuseDashModToolsUI.Models;
 
-public class Settings : ISettings
+public class Setting
 {
     public string? MuseDashFolder { get; set; }
+    [JsonIgnore] public string ModsFolder => !string.IsNullOrEmpty(MuseDashFolder) ? Path.Join(MuseDashFolder, "Mods") : string.Empty;
     public AskType AskInstallMuseDashModTools { get; set; } = AskType.Always;
     public AskType AskEnableDependenciesWhenInstalling { get; set; } = AskType.Always;
     public AskType AskDisableDependenciesWhenDeleting { get; set; } = AskType.Always;
