+++ UI/Views/DownloadWindow.axaml
@@ -1,47 +1,38 @@
 <Window xmlns="https://github.com/avaloniaui"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:vm="using:MuseDashModToolsUI.ViewModels"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:controls="clr-namespace:Material.Styles.Controls;assembly=Material.Styles"
-        mc:Ignorable="d" d:DesignWidth="400" d:DesignHeight="250"
         x:Class="MuseDashModToolsUI.Views.DownloadWindow"
         x:DataType="vm:DownloadWindowViewModel"
         x:CompileBindings="True"
         Title="DownloadWindow"
-        Width="400"
-        Height="250">
+        Width="320">
 
-    <Grid RowDefinitions="*,*,50">
+    <Grid RowDefinitions="120,100">
         <Image Grid.Row="0"
                VerticalAlignment="Top"
+               Margin="0,20,0,0"
                Width="50"
                Source="/Assets/ML_Icon.png" />
         <Border Grid.Row="0"
                 VerticalAlignment="Center"
-                Margin="0,20,0,0">
+                Margin="0,60,0,10">
             <TextBlock Classes="Headline5"
                        TextAlignment="Center"
                        Foreground="#6D96FF"
                        Text="Downloading MelonLoader" />
         </Border>
 
-        <controls:Card Grid.Row="1" Width="250" Height="80">
+        <controls:Card Grid.Row="1" Width="250" Height="80" >
             <StackPanel VerticalAlignment="Center">
                 <ProgressBar Margin="20,20,20,0"
                              Value="{Binding Percentage}" />
-                <TextBlock Text="{Binding DownloadProgress} " 
+                <TextBlock Text="DownLoad Progress: 0%" 
                            TextAlignment="Center"
                            Margin="0,10,0,0"/>
 
             </StackPanel>
         </controls:Card>
-        <Button Grid.Row="2"
-                IsEnabled="{Binding DownloadFinished}"
-                Classes="Dark"
-                Margin="5"
-                HorizontalAlignment="Right"
-                Content="Finished" />
     </Grid>
 
 </Window>
